Move gem files to properly named folder structure
